+++ Controller/Adminhtml/Install/Run.php
@@ -49,7 +49,7 @@ class Run extends \Autocompleteplus\Autosuggest\Controller\Adminhtml\Install
                 $this->api->sendError('Could not get license string.');
                 $this->messageManager->addError(__('Something went wrong when trying to install InstantSearch+'));
                 $result->setPath('adminhtml/dashboard/index');
-                return $this;
+                return $result;
             }
 
             $this->api->setApiUUID($responseData->uuid);
@@ -58,8 +58,8 @@ class Run extends \Autocompleteplus\Autosuggest\Controller\Adminhtml\Install
         } else {
             $this->messageManager->addError(__('Something went wrong when trying to install InstantSearch+'));
         }
-        //$this->_redirect('adminhtml/dashboard/index');
         $result->setPath('adminhtml/dashboard/index');
+
         return $result;
     }
 }