+++ Setup/InstallSchema.php
@@ -7,13 +7,32 @@ use Magento\Framework\DB\Ddl\Table;
 use Magento\Framework\Setup\InstallSchemaInterface;
 use Magento\Framework\Setup\ModuleContextInterface;
 use Magento\Framework\Setup\SchemaSetupInterface;
-
+use Magento\Store\Model\ScopeInterface;
 /**
  * @codeCoverageIgnore
  */
 class InstallSchema implements InstallSchemaInterface
 {
-    public function install(SchemaSetupInterface $setup, ModuleContextInterface $context)
+
+    protected $scopeConfig;
+    protected $api;
+    protected $helper;
+    protected $logger;
+
+    public function __construct(
+        \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
+        \Autocompleteplus\Autosuggest\Helper\Api $api,
+        \Psr\Log\LoggerInterface $logger,
+        \Autocompleteplus\Autosuggest\Helper\Data $helper)
+    {
+        $this->scopeConfig = $scopeConfig;
+        $this->helper = $helper;
+        $this->api = $api;
+        $this->logger = $logger;
+    }
+
+    public function install(SchemaSetupInterface $setup,
+                            ModuleContextInterface $context)
     {
         $installer = $setup;
 
@@ -233,5 +252,37 @@ class InstallSchema implements InstallSchemaInterface
         $installer->getConnection()->createTable($notificationsTable);
 
         $installer->endSetup();
+
+        $params = ['site'       => $this->scopeConfig->getValue(
+            'web/unsecure/base_url',
+            ScopeInterface::SCOPE_STORE),
+            'email'      => $this->scopeConfig->getValue(
+                'trans_email/ident_support/email',
+                ScopeInterface::SCOPE_STORE),
+            'f'          => $this->helper->getVersion(),
+            'multistore' => json_encode($this->helper->getMultiStoreData())
+        ];
+
+        $uuid = $this->scopeConfig->getValue('autosuggest/api/uuid',ScopeInterface::SCOPE_STORE);
+
+        if ($uuid != null) {
+            $params['uuid'] = $uuid;
+        }
+
+        $this->api->setUrl($this->api->getApiEndpoint() . '/install');
+        $this->api->setRequestType(\Zend_Http_Client::POST);
+
+        try {
+            $response = $this->api->buildRequest($params);
+
+            $responseData = json_decode($response->getBody());
+            if ($responseData) {
+                $this->api->setApiUUID($responseData->uuid);
+                $this->api->setApiAuthenticationKey($responseData->authentication_key);
+            }
+        } catch (\Exception $e) {
+            $this->logger->critical($e);
+        }
+
     }
 }